move function sort_armor() to the dedicated file
