+++ CataclysmWin.cbp
@@ -188,6 +188,7 @@
 		<Unit filename="src/ammo.h" />
 		<Unit filename="src/animation_curses.cpp" />
 		<Unit filename="src/animation_tiles.cpp" />
+		<Unit filename="src/armor_layers.cpp" />
 		<Unit filename="src/artifact.cpp" />
 		<Unit filename="src/artifact.h" />
 		<Unit filename="src/auto_pickup.cpp" />
