+++ msvc100/Cataclysm.vcxproj.filters
@@ -385,6 +385,9 @@
     <ClCompile Include="..\src\animation_tiles.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\armor_layers.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\artifact.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
