+++ msvc110/Cataclysm.vcxproj
@@ -89,6 +89,7 @@
     <ClCompile Include="..\src\ammo.cpp" />
     <ClCompile Include="..\src\animation_curses.cpp" />
     <ClCompile Include="..\src\animation_tiles.cpp" />
+    <ClCompile Include="..\src\armor_layers.cpp" />
     <ClCompile Include="..\src\artifact.cpp" />
     <ClCompile Include="..\src\auto_pickup.cpp" />
     <ClCompile Include="..\src\basecamp.cpp" />
