+++ msvc110/Cataclysm.vcxproj.filters
@@ -402,6 +402,9 @@
     <ClCompile Include="..\src\mondefense.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\armor_layers.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\action.h">
